fix(sandbox): close CA cache-TTL and authorize-deadline gaps
Third review round on PR #6723:
- ca.rs: cache TTL now tracks each leaf's actual (root-capped) not_after
  instead of the raw leaf_ttl, so a leaf minted near the CA's own expiry
  can no longer be served from cache after its certificate has expired.
  Adds a test exercising the root-expiry cap directly.
- credential_firewall.rs: authorize() rechecks its deadline once more
  after reclaiming expired entries and before cloning the live set, so a
  Grant can never be returned once the deadline has passed even if that
  in-lock work is what crossed it.
- credential_firewall.rs: Debug now uses try_lock (mirroring
  RuntimeSecretInjectionStore) instead of an unconditional lock, so
  formatting can never block on the same mutex stage/revoke/authorize
  hold.
- ca.rs: split the 943-line test module into ca/tests.rs, matching the
  precedent already set for credential_firewall.rs.
